--- src/main/java/edu/millersville/uml_editor/Relationships.java
@@ -1,70 +1,0 @@
-package edu.millersville.uml_editor;
-
-
-import java.util.*;
-
-public class Relationships {
-
-	private ClassObject source;
-	private ClassObject destination;
-	private static String ID;
-	private String type;
-
-///////////////////////////////////////////////////////////
-//
-//	Constructor
-//
-///////////////////////////////////////////////////////////
-
-	public Relationships(ClassObject class1, ClassObject class2, String id, String type1) 
-	{
-		source = class1;
-		destination = class2;
-		ID = id;
-		type = type1;
-	}
-	
-///////////////////////////////////////////////////////////
-//
-//	sourceName
-//
-///////////////////////////////////////////////////////////
-
-	public String sourceName()
-	{
-		return source.getClassName();
-	}
-	
-///////////////////////////////////////////////////////////
-//
-//	destinatinoName
-//
-///////////////////////////////////////////////////////////
-
-	public String destinationName()
-	{
-		return destination.getClassName();
-	}
-
-///////////////////////////////////////////////////////////
-//
-//	changeType
-//
-///////////////////////////////////////////////////////////
-
-	public void changeType(String newType)
-	{
-		type = newType;
-	}
-
-///////////////////////////////////////////////////////////
-//
-//	relType
-//
-///////////////////////////////////////////////////////////
-
-	public String relType()
-	{
-		return type;
-	}
-}
