+++ src/main/java/edu/millersville/uml_editor/view/package-info.java
@@ -1,0 +1,8 @@
+/**
+ * 
+ */
+/**
+ * @author yangc
+ *
+ */
+package edu.millersville.uml_editor.view;
